fixed some errors

+++ print_numbers.c
@@ -9,10 +9,10 @@
 int print_d(va_list d)
 {
 	int j, count;
-	long long int n;
+	long int n;
 	int *a;
 	int size = 0;
-	n = va_arg(d, long long int);
+	n = va_arg(d, long int);
 	count = 0;
 
 	if (n == 0)
